requirements.txt
Django==5.0
djangorestframework==23.3.2
django-rest-auth==23.3.2
django-allauth==23.3.2
Pillow==10.1.0
django-environ==0.11.2
psycopg2-binary==2.9.9
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
inflect==7.0.0
django-crispy-forms==2.1
python-dotenv==1.0.0
django_axes==6.3.0
